requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
scikit-learn==1.3.1
Pillow==10.0.1
sklearn_extras==0.1.1